wip: Start using a PachliAccountId sealed type
The account ID can either be -1 (the active account) or a specific
account ID.

It's currently possible to send -1 to functions that don't expect it,
which can cause a crash, or operations not to happen (because no entity
in the database has that ID).

Fix this by making the account ID a sealed interface that can represent
these two states, and be explicit in function parameters as to whether
or not the function can accept the active account and a specific account
or just a specific account.