fix: Move the "Toot" button to top right (#2034)
Adding the "quote visibility" button meant there isn't enough space on
small-width devices to show the "Toot" button on the bottom row when
composing.

To fix this, move the "Toot" button to the right of the toolbar,
matching the UI guidelines.

Move the other buttons in the toolbar to the left.

Add the user's avatar image as a compound drawable to the "Posting as"
`TextView`.